fix(mm-next/video_category): insert GPT Ads `MB_FT`&`PC_FT` into videos list
